src/services/requirements.txt
cassandra_driver==3.25.0
fastapi==0.115.12
Flask==3.1.0
flask_cors==5.0.1
httpx==0.28.1
pydantic==2.11.4
pymongo==4.12.1
python-dotenv==1.1.0
redis>=4.5.0,<6.0.0
Requests==2.32.3
rq==2.3.3
uvicorn==0.34.2
